% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_locs.R
\name{find_locs}
\alias{find_locs}
\title{Find locations of stations/etc. from spatial inputs}
\usage{
find_locs(x, lat = NULL, lon = NULL, radius = NULL, bbox = NULL, ...)
}
\arguments{
\item{lat, lon}{(numeric) Latitude and longitude, in decimal degree. One or
more. The lat and lon vectors must be the same length.}

\item{radius}{(numeric) Radius (in km) to search from the lat,lon coordinates}

\item{bbox}{(numeric) Bounding box, of the form: min-longitude, min-latitude,
max-longitude, max-latitude}
}
\description{
Find locations of stations/etc. from spatial inputs
}
\examples{
\dontrun{
# Single point pair
x <- rnoaa::isd_stations()
find_locs(x, lat = 40, lon = -120, radius = 50)

# Many point pairs
x <- rnoaa::isd_stations()
find_locs(x, lat = c(30, 40), lon = c(-120, -120), radius = 50)
}
}

