% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sp_extract_gridded.R
\name{sp_extract_gridded}
\alias{sp_extract_gridded}
\title{Extract environmental data - gridded data based}
\usage{
sp_extract_gridded(x, from = "noaa_sst", radius = 50, select = "first",
  date = NULL)
}
\arguments{
\item{x}{input data.frame}
}
\description{
Extract environmental data - gridded data based
}
\details{
Works for the use case of finding locations for point based stations,
floats/buoys type data
}
\examples{
\dontrun{
file <- system.file("examples", "obis_mola_mola.csv", package = "spenv")
dat <- read.csv(file)
head(dat)
res <- sp_extract_gridded(x = dat[1:10,], radius = 20)
}
}

