% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_extract_pt.R
\name{sp_extract_pt}
\alias{sp_extract_pt}
\title{Extract environmental data - pt based}
\usage{
sp_extract_pt(x, from = "noaa_isd", radius = 50, select = "first",
  date = NULL)
}
\arguments{
\item{x}{input data.frame}

\item{from}{which data source, only 'noaa_isd' for now}

\item{radius}{radius, in km}

\item{select}{which}

\item{date}{Date}
}
\description{
Extract environmental data - pt based
}
\details{
Works for the use case of finding locations for point based stations,
floats/buoys type data
}
\examples{
\dontrun{
file <- system.file("examples", "obis_mola_mola.csv", package = "spenv")
dat <- read.csv(file)
head(dat)
res <- sp_extract_pt(x = dat[1:10,], radius = 20)
}
}
\author{
Scott Chamberlain
}

