% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_mutate.R
\name{sp_mutate}
\alias{sp_mutate}
\title{code}
\usage{
sp_mutate(x, from = "noaa_isd", radius = 50, select = "first",
  date = NULL)
}
\arguments{
\item{x}{input data.frame}
}
\description{
code
}
\examples{
\dontrun{
file <- system.file("examples", "obis_mola_mola.csv", package = "spenv")
dat <- read.csv(file)
head(dat)

res <- sp_mutate(x = dat[1:10,], radius = 100)

}
}

