% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_query.R
\name{sp_query}
\alias{sp_query}
\title{Explore what data is available}
\usage{
sp_query(query)
}
\arguments{
\item{query}{Query terms}
}
\description{
Explore what data is available
}
\examples{
\dontrun{
sp_query(query = "precipitation")
sp_query("temperature")
}
}

