% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spenv-package.R
\docType{package}
\name{spenv-package}
\alias{spenv}
\alias{spenv-package}
\title{Spatial data to environmental data extractor}
\description{
Spatial data to environmental data extractor
}
\keyword{package}

