% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sp_query.R
\name{sp_query}
\alias{sp_query}
\title{Explore what data is available}
\usage{
sp_query(variable = NULL, source = NULL)
}
\arguments{
\item{variable}{Query for a variable}

\item{source}{Query for a data source}
}
\description{
Explore what data is available
}
\examples{
\dontrun{
# THESE DON'T ACTUALLY WORK YET...
# Query by data variable
sp_query(variable = "precipitation")
sp_query(variable = "temperature")

# Query by data variable
sp_query(source = "noaa")

# Query by spatial resolution
sp_query(spatial_res = "noaa")

# Query by temporal resolution
sp_query(temporal_res = "weekly")

# Query by spatial coverage
sp_query(spatial_coverage = "global")

# Query by temporal coverage
sp_query(temporal_coverage = "1950:1970")

# Query by license
sp_query(license = "CC0")
}
}

